Assume Dropbox is in the user's home folder and get that path programatically
The program will assume that Dropbox is in /Users/blah/Dropbox (mac)
or c:\Users\blah\Dropbox (windows) and get that path programatically.

The only thing that needs to be changed is the dashboard folder (ie. which
folder to use within dropbox). I have that hardcoded now to be
YL_tasks/yl_pilot