Fixed Star Tutorial issue of the stars not resetting
The issue was that the practice was actually not using the stars shown
in the first tutorial, but a bunch of random ones, so when it went
through the tutorial again the program would not update these stars to
hide. Also it seemed to not properly reshow the second tutorial's
objects. Updated the program so it will use the same set of stars
through all the practices/tutorials.