Update csv file to have lastcompleted column instead of completed and update the grayed out lines in the player selector to not be selectable
Updated the grayed out player names in the config's player selector so
there colliders would be turned off so they won't be selectable. Also
updated the csv complete column to now be lastcompleted which will
contain the int of the last game played. Now when a player is signed
out before finishing all the tasks, it will save the last task that was
completed so when the player gets resigned in it will start off at this
point.